Make some parts pub(crate)
